pyproject.toml
[tool.poetry]
name = "segmentation"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
clip = {git = "https://github.com/openai/CLIP.git"}
ftfy = "^6.1.3"
iprogress = "^0.4"
ipykernel = "^6.29.2"
kaggle = "^1.6.3"
lightning = "^2.1.3"
matplotlib = "^3.8.2"
nbqa = "^1.8.3"
numpy = "^1.26.3"
open-clip-torch = "^2.24.0"
opencv-python = "^4.9.0.80"
optuna = "^3.6.1"
optuna-dashboard = "^0.15.1"
prettytable = "^3.10.0"
python = "^3.10"
pytorch-lightning = "^2.1.3"
regex = "^2023.12.25"
requests = "^2.31.0"
scikit-image = "^0.22.0"
scikit-learn = "^1.4.2"
scipy = "^1.12.0"
seaborn = "^0.13.1"
segment-anything = {git = "https://github.com/facebookresearch/segment-anything.git"}
tensorboard = "^2.15.1"
torch = "^2.1.2"
torchcam = "^0.4.0"
torchvision = "^0.16.2"
tqdm = "^4.66.2"
xgboost = "^2.0.3"
umap-learn = "^0.5.6"
types-requests = "^2.31.0.20240406"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.4"
flake8 = "^7.0.0"
isort = "^5.13.2"
mypy = "^1.8.0"
nbqa = {extras = ["toolchain"], version = "^1.8.5"}
pandas-stubs = "^2.2.1.240316"
pre-commit = "^3.6.2"
pytest = "^8.0.2"
pytest-cov = "^4.1.0"
ruff = "^0.3.1"
types-pillow = "^10.2.0.20240213"
types-tqdm = "^4.66.0.20240106"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
